@@ -9,53 +9,57 @@
   },
   "private": true,
   "dependencies": {
-    "@emotion/react": "11.9.0",
-    "@emotion/styled": "11.8.1",
+    "@emotion/react": "11.10.6",
+    "@emotion/styled": "11.10.6",
     "core-js": "^3.22.5",
     "date-fns": "^2.28.0",
     "firebase": "^9.8.1",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
-    "react-router-dom": "^6.3.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.11.2",
     "regenerator-runtime": "0.13.7",
     "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@emotion/babel-plugin": "11.7.2",
+    "@emotion/babel-plugin": "11.10.6",
     "@faker-js/faker": "^7.3.0",
-    "@nrwl/cli": "13.10.2",
-    "@nrwl/cypress": "13.10.2",
-    "@nrwl/eslint-plugin-nx": "13.10.2",
-    "@nrwl/jest": "13.10.2",
-    "@nrwl/linter": "13.10.2",
-    "@nrwl/node": "^13.10.4",
-    "@nrwl/react": "^13.10.4",
-    "@nrwl/web": "^13.10.4",
-    "@nrwl/workspace": "13.10.2",
+    "@nx/cypress": "16.3.1",
+    "@nx/eslint-plugin": "16.3.1",
+    "@nx/jest": "16.3.1",
+    "@nx/js": "^16.3.1",
+    "@nx/linter": "16.3.1",
+    "@nx/node": "16.3.1",
+    "@nx/react": "16.3.1",
+    "@nx/web": "16.3.1",
+    "@nx/webpack": "^16.3.1",
+    "@nx/workspace": "16.3.1",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "13.0.0",
-    "@testing-library/react-hooks": "7.0.2",
+    "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "^14.3.0",
-    "@types/jest": "27.0.2",
-    "@types/node": "16.11.7",
-    "@types/react": "18.0.0",
-    "@types/react-dom": "18.0.0",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
-    "babel-jest": "27.2.3",
+    "@types/jest": "29.4.4",
+    "@types/node": "18.14.2",
+    "@types/react": "18.0.28",
+    "@types/react-dom": "18.0.11",
+    "@typescript-eslint/eslint-plugin": "5.59.8",
+    "@typescript-eslint/parser": "5.59.8",
+    "babel-jest": "29.4.3",
     "cypress": "^9.7.0",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.4.0",
-    "jest": "27.2.3",
-    "nx": "13.10.2",
-    "prettier": "^2.5.1",
-    "react-test-renderer": "18.0.0",
-    "ts-jest": "27.0.5",
-    "typescript": "^4.6.4"
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-react": "7.32.2",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "jest": "29.4.3",
+    "jest-environment-jsdom": "28.1.3",
+    "nx": "16.3.1",
+    "nx-cloud": "16.0.5",
+    "prettier": "2.8.8",
+    "react-refresh": "^0.14.0",
+    "ts-jest": "29.1.0",
+    "ts-node": "10.9.1",
+    "typescript": "5.0.4"
   }
 }
